feat: add empty state to empty dashboard on no edit mode
