+++ packages/frontend/src/components/DashboardTiles/EmptyStateNoTiles/EmptyStateNoTiles.styles.tsx
@@ -1,0 +1,10 @@
+import { Colors, Icon } from '@blueprintjs/core';
+import styled from 'styled-components';
+
+export const NoSavedChartsWrapper = styled.div``;
+
+export const EmptyStateIcon = styled(Icon)`
+    path {
+        fill: ${Colors.GRAY3};
+    }
+`;
